index on bugfix-12075: 17ebae3 [[ ExecValue-ization ]] Reverts update on the literal numbers - using them as proper numbers when possible breaks the formatting of floats
